+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:guide_for_hajj/test.dart';
+import 'package:guide_for_hajj/guide_hagg_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -8,13 +8,13 @@ void main() {
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
+  
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Guide For Hajj',
-      home: RotatingCircle(),
+      home: HajjGuide(),
     );
   }
 }
