+++ pubspec.yaml
@@ -61,7 +61,6 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/images/frame.svg
     - assets/images/
 
   # An image asset can refer to one or more resolution-specific "variants", see
